small changes in Rechenaufgabe
